requirements.txt
aiohttp==3.9.3
aiosignal==1.3.2
appnope==0.1.4
asttokens==3.0.0
asyncio==3.4.3
attrs==24.3.0
aws-lambda-powertools==1.25.5
aws-xray-sdk==2.14.0
beautifulsoup4==4.12.3
behave==1.2.6
boto3==1.35.84
botocore==1.35.84
certifi==2024.12.14
chardet==5.2.0
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.2
cssselect==1.2.0
debugpy==1.8.11
decorator==5.1.1
Deprecated==1.2.15
exceptiongroup==1.2.2
execnet==2.1.1
executing==2.1.0
fasteners==0.19
fastjsonschema==2.21.1
filelock==3.16.1
frozenlist==1.5.0
fyers_apiv3==3.1.4
h11==0.14.0
h5py==3.12.1
idna==3.7
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.30.0
jedi==0.19.2
jmespath==0.10.0
jupyter_client==8.6.3
jupyter_core==5.7.2
markdown-it-py==3.0.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mss==10.0.0
multidict==6.1.0
nest-asyncio==1.6.0
nodriver==0.38.post1
numpy==2.2.0
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
parameterized==0.9.0
parse==1.20.2
parse_type==0.6.4
parso==0.8.4
pdbp==1.5.0
pexpect==4.9.0
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.48
propcache==0.2.1
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
py==1.11.0
Pygments==2.18.0
pynose==1.5.1
pyotp==2.9.0
PySocks==1.7.1
pytest==8.2.1
pytest-html==2.0.1
pytest-metadata==3.1.1
pytest-ordering==0.6
pytest-rerunfailures==14.0
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
redis==5.2.1
requests==2.31.0
rich==13.7.1
s3transfer==0.10.4
sbvirtualdisplay==1.3.0
selenium==4.21.0
seleniumbase==4.27.5
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
tabcompleter==1.3.0
tornado==6.4.2
traitlets==5.14.3
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
websocket-client==1.6.1
websockets==13.1
wrapt==1.17.0
wsproto==1.2.0
yarl==1.18.3
